Changed base
Int base is now a double base!

+++ main.cpp
@@ -2,21 +2,21 @@
 
 using namespace std;
 
-void power(int base, int exp);
+void power(double base, int exp);
 
 int main()
 {
     cout<<"GitHub Test"<<endl;
     
-    power(10, 2);
-    power(5, 3);
-    power(10, 1);
-    power(10, 0);
+    power(10.0, 2);
+    power(5.0, 3);
+    power(10.0, 1);
+    power(10.0, 0);
 
     return 0;
 }
 //--
-void power(int base, int exp)
+void power(double base, int exp)
 {
     int result = 1;
     for(int i = 0;i < exp;i++)
